Restructure the object.
Constructor now merely validates the taxonomy and stores it "as is" as an
array of array.  It no longer stores in itself the results of certain queries
on itself or alternate methods of reporting the data.  Those will now be
handled by discrete methods.  INCOMPLETE; NOT STABLE; SOME TESTS BREAKING.

Rename two existing test files and add new file t/002-getters.t  Update
MANIFEST and .gitignore.