Introduce attribute 'row_analysis'.
We will use this KVP within the object to record data about each node, keyed
on the node's path.  We will not directly expose this attribute to the user
but will, as needed, provide accessors.

Subsume 'row_depth' as an attribute of each node's entry in 'row_analysis'.
Rework 'descendant_counts()' to use this new structure.