Upon entering nestify(), we must make sure that row_analysis is clean.
It's not clear where, in the course of nestify(), $self->{row_analysis} was
getting assigned to.  Wherever it was, it was associated with non-idempotent
and incorrect results when calling nestify() on the same object a second time.
Reducing $self->{row_analysis} to just holding row_depth appears to resolve
the problem.